Made Sandstorm sound good
I found a better voicegroup for it than I had been using, but I made a
change to one of the instruments: it had been a channel 1 square wave,
but I needed a channel 2 square wave because the cursor sound effect was
drowning out the lead in the song. I also tweaked volumes, found better
instruments for the bass, and removed some unnecessary channels.